// src/app/Models/ProyectosModels.ts
export class ProyectosModels {

    MatriculaInmobiliariaProyecto: number;
    NombreProyecto: string;
    DireccionProyecto: string;
    EstratoProyecto: number;
    EscrituraReglamentoProyecto: string;
    AdministradorProyecto: string;
    TelefonoAdministradorProyecto: string;
    CorreoAdministradorProyecto: string;
}